SMI-S Plugin: Allow provision parameter of volume_create()
 * Allow 'provision' parameter of volume_create() to be:
    Volume.PROVISION_DEFAULT
    Volume.PROVISION_FULL
    Volume.PROVISION_THIN

 * For Volume.PROVISION_DEFAULT, we create full allocated volume
   unless only thin volume is supported. [1]

 * For full allocated volume, we set 'ElementType' of
   CreateOrModifyElementFromStoragePool() to dmtf.ELEMENT_THICK_VOLUME(2)

 * For thin provisioning volume, we set 'ElementType' of
   CreateOrModifyElementFromStoragePool() to dmtf.ELEMENT_THIN_VOLUME(5)

 * This also fix a bug when running plugin_test.py against HDS AMS where
   only dmtf.ELEMENT_THIN_VOLUME(5) is supported.

[1] HDS AMS Thin Pool only support ThinlyProvisionedStorageVolume.

+++ plugin/smispy/dmtf.py
@@ -93,8 +93,8 @@ SUPPORT_ELEMENT_REDUCE = 13
 
 # DMTF CIM 2.37.0 experimental CIM_StorageConfigurationCapabilities
 # ['SupportedStorageElementTypes']
-ELEMENT_THICK_VOLUME = 2
-ELEMENT_THIN_VOLUME = 5
+ELEMENT_THICK_VOLUME = Uint16(2)
+ELEMENT_THIN_VOLUME = Uint16(5)
 
 # DMTF CIM 2.29.1 CIM_StorageConfigurationCapabilities
 # ['SupportedStoragePoolFeatures']
